folder and files reorganization (MovieTilesContainer folder added in common/Tiles)
