+++ src/features/MainPage/styled.js
@@ -11,42 +11,6 @@ export const Container = styled.div`
   }
 `;
 
-export const Title = styled.h1`
-  font-size: 36px;
-  font-weight: 600;
-  line-height: 43.2px;
-  margin-top: 56px;
-  margin-bottom: 24px;
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.mobile}) {
-    font-size: 20px;
-    line-height: 24px;
-    margin-top: 24px;
-    margin-bottom: 12px;
-  }
-`;
-
-export const TilesContainer = styled.div`
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-gap: 24px;
-  margin-bottom: 40px;
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.tablet}) {
-    grid-template-columns: repeat(3, 1fr);
-  }
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.mobile}) {
-    grid-template-columns: repeat(2, 1fr);
-    grid-gap: 16px;
-    margin-bottom: 32px;
-  }
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.largeMobile}) {
-    grid-template-columns: 1fr;
-  }
-`;
-
 export const LoadingPage = styled.div``;
 
 export const ErrorPage = styled.div``;
